Create README.md
[3/28/2020] 
Time resolved data on cases and deaths in all Pennsylvania (PA) countries since 3/19/2020 until 3/29/2020 were scraped from PA Dept of Health and Allegheny County's data (i.e. the county for the City of Pittsburgh) was extracted for modeling. These data were fed into a logistic regression model using only the data available i.e. 10 days prior to the date of analysis. The choice of a logistic fit is predicated by the trends in case volumes seen in cities in China, where social distancing and best practices have resulted in some success in curbing the spread of COVID-19.
The Greater Pittsburgh Area is has an estimated capacity of between 398 and 500 ICU beds and about 10x more hospital beds. About 7-10% of cases are expected to lead to hospitalization with the 19% of Allegheny County's population over 65 years of age being at the highest risk of hospitalization related deaths. At a 7% ICU hospitalization rate, Allegheny County can tolerate roughly 6500 cases of COVID-19. This doesn't factor in ICU transfers and patient movement out of ICUs owing to other reasons.
To hit 6500 cases at the current case growth rates would require significant lapses in social distancing best practices or an otherwise unforseen spike in case volumes.
However, if cases were to accumulate at the rate that it is across the US as a whole (i.e. case volume doubles every 3 days, based on data at the present time), a peak capacity situation may develop on the 15th day from 3/28/2020 (the date of this analysis).

Image "3-28-2020--Allegheny.PNG" is the fit referenced in the paragraphs above.