SIRD model prediction
3/30/2020:
My best model yet for Allegheny County - this factors in the following ODEs to solve for susceptible, infected, recovered and dead cases for COVID19, based on the 10 days of historical data available between 3/19 and 3/29 ... This model forecasts susceptible individual counts, recovered individual counts and deaths, assuming currently that the "current" recovery rate is 20% of the susceptible individual counts. If recovery rates baseline to higher than 20%, then the numbers can be more optimistic!  The latter is consistent with the median recovery rate from Johns Hopkin's data by country.